+++ src/game/scripts/vscripts/commands.lua
@@ -71,6 +71,17 @@ function Commands:OnPlayerChat(keys)
         elseif OptionManager:GetOption('neutralMultiply') > 1 then
             util:DisplayError(playerID, "#multiplyAlreadyOn")
         end
+    elseif IsCommand("-universalshops") or IsCommand("-us") then
+       if OptionManager:GetOption('universalShops') == 0 then
+            util:CreateVoting("lodVotingUniversalShops", playerID, 10, OptionManager:GetOption('mapname') == 'all_allowed' and 50 or 100, function()
+                OptionManager:SetOption('universalShops', 1)
+                GameRules:SetUseUniversalShopMode(true)
+                EmitGlobalSound("Event.CheatEnabled")
+                GameRules:SendCustomMessage('Enough players voted to enable universal shops. <font color=\'#70EA72\'>You can now buy any item from any shop</font>.',0,0)
+            end)
+        else
+            util:DisplayError(playerID, "#universalShopsAlreadyOn")
+        end
     elseif IsCommand("-enablecheat") or IsCommand("-ec") then
         if not ingame.voteEnabledCheatMode and not Convars:GetBool("sv_cheats") then
             util:CreateVoting("lodVotingEnableCheatMode", playerID, 10, 100, function()