+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 - Cooldown and Spell Amp survivor now require you to be near an enemy hero
 - Removed cooldown reduction talents
 - Fixed Astral Imprison not doing damage
+- Hopefully fixed AA perk
+*
+- Glaives Int Steal is now ultimate ability
+- You can now vote to enable universal shops with -us or -universalshops or from the vote menu. Requires 50% to succeed
 
 ## 9.2
 - Astral Spirit Return will no longer trigger Essence Aura